requirements.txt
Django==4.0.4
djangorestframework==3.13.1
environ==1.0
psycopg2-binary==2.9.3
Pillow==9.1.1
djangorestframework-simplejwt==5.1.0
django-filter==21.1
django-environ==0.8.1
coverage==6.4
selenium==4.2.0
beautifulsoup4==4.11.1
requests==2.27.1
lxml==4.9.0
ipdb==0.13.9
webdriver-manager==3.7.0
celery==5.2.7
redis==4.3.3
flower==1.0.0
django-celery-beat==2.3.0
django-elasticsearch==7.0.2
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
